cargo doc generated documentation for blending module
